[changeset] extract pipe_registry to a separate file
